requirements.txt
crewai
langchain-groq
prometheus-api-client
requests
streamlit
fpdf
docker
flask
prometheus_client
python-dotenv
fastapi
uvicorn[standard]
pydantic
